Fix pagination bug
Fixed pagination bug that occurs if the last item in the previous batch
of paginated and the actual first item in the next batch have the same
value in the property the documents are being sorted on. The correct first item
in the next batch won't be returned since the new documents are first checked
if their filter property is greater than the last item's filter property.